# app/blueprints/googlegeminimodels_bp.py
# GENERATED BY ARCHIE — entity: googlegeminimodels
# solution_id: 5373
# bundle_id: code_2026-04-06T18-28-31Z
# DO NOT EDIT IN PLACE -- REGENERATE FROM ARCHITECTURE

from __future__ import annotations

import uuid
import logging
from http import HTTPStatus

from flask import Blueprint, jsonify, request
from flask_jwt_extended import jwt_required, get_jwt_identity
from marshmallow import ValidationError
from sqlalchemy import or_

from database import db
from app.models.googlegeminimodels import Googlegeminimodels
from app.schemas.googlegeminimodels_schema import (
    GooglegeminimodelsSchema,
    GooglegeminimodelsCreateSchema,
    GooglegeminimodelsUpdateSchema,
)

logger = logging.getLogger(__name__)

googlegeminimodels_bp = Blueprint(
    "googlegeminimodels",
    __name__,
)

_schema = GooglegeminimodelsSchema()
_schema_many = GooglegeminimodelsSchema(many=True)
_create_schema = GooglegeminimodelsCreateSchema()
_update_schema = GooglegeminimodelsUpdateSchema()

# ---------------------------------------------------------------------------
# GET /  — list with pagination + optional search
# ---------------------------------------------------------------------------

@googlegeminimodels_bp.route("/", methods=["GET"])
@jwt_required()
def list_googlegeminimodelss():
    """Return a paginated list of googlegeminimodels records."""
    page = request.args.get("page", 1, type=int)
    per_page = min(request.args.get("per_page", 20, type=int), 100)
    search = request.args.get("search", "", type=str).strip()

    query = Googlegeminimodels.query

    paginated = query.order_by(Googlegeminimodels.created_at.desc()).paginate(
        page=page, per_page=per_page, error_out=False
    )
    return jsonify(
        {
            "items": _schema_many.dump(paginated.items),
            "total": paginated.total,
            "page": paginated.page,
            "pages": paginated.pages,
            "per_page": paginated.per_page,
        }
    )

# ---------------------------------------------------------------------------
# GET /<id>  — single record
# ---------------------------------------------------------------------------

@googlegeminimodels_bp.route("/<string:item_id>", methods=["GET"])
@jwt_required()
def get_googlegeminimodels(item_id: str):
    item = db.session.get(Googlegeminimodels, item_id)
    if item is None:
        return jsonify({"error": "not found"}), HTTPStatus.NOT_FOUND
    return jsonify(_schema.dump(item))

# ---------------------------------------------------------------------------
# POST /  — create
# ---------------------------------------------------------------------------

@googlegeminimodels_bp.route("/", methods=["POST"])
@jwt_required()
def create_googlegeminimodels():
    try:
        data = _create_schema.load(request.get_json(force=True) or {})
    except ValidationError as exc:
        return jsonify({"errors": exc.messages}), HTTPStatus.UNPROCESSABLE_ENTITY

    item = Googlegeminimodels(
        id=str(uuid.uuid4()),

        **{k: v for k, v in data.items() if v is not None},
    )
    db.session.add(item)
    db.session.commit()
    logger.info("Created googlegeminimodels id=%s user=%s", item.id, get_jwt_identity())
    return jsonify(_schema.dump(item)), HTTPStatus.CREATED

# ---------------------------------------------------------------------------
# PUT /<id>  — full update
# ---------------------------------------------------------------------------

@googlegeminimodels_bp.route("/<string:item_id>", methods=["PUT"])
@jwt_required()
def update_googlegeminimodels(item_id: str):
    item = db.session.get(Googlegeminimodels, item_id)
    if item is None:
        return jsonify({"error": "not found"}), HTTPStatus.NOT_FOUND

    try:
        data = _update_schema.load(request.get_json(force=True) or {})
    except ValidationError as exc:
        return jsonify({"errors": exc.messages}), HTTPStatus.UNPROCESSABLE_ENTITY

    for key, value in data.items():
        if value is not None:
            setattr(item, key, value)

    db.session.commit()
    logger.info("Updated googlegeminimodels id=%s user=%s", item.id, get_jwt_identity())
    return jsonify(_schema.dump(item))

# ---------------------------------------------------------------------------
# DELETE /<id>
# ---------------------------------------------------------------------------

@googlegeminimodels_bp.route("/<string:item_id>", methods=["DELETE"])
@jwt_required()
def delete_googlegeminimodels(item_id: str):
    item = db.session.get(Googlegeminimodels, item_id)
    if item is None:
        return jsonify({"error": "not found"}), HTTPStatus.NOT_FOUND
    db.session.delete(item)
    db.session.commit()
    logger.info("Deleted googlegeminimodels id=%s user=%s", item_id, get_jwt_identity())
    return "", HTTPStatus.NO_CONTENT