# openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Archiet \u2014 AI-Native Architecture-to-Code SaaS Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/google_gemini_models": {
      "get": {
        "operationId": "list_google_gemini_modelss",
        "summary": "List GoogleGeminiModelss",
        "tags": [
          "google_gemini_models"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-archimate-source": 1133
      },
      "post": {
        "operationId": "create_google_gemini_models",
        "summary": "Create GoogleGeminiModels",
        "tags": [
          "google_gemini_models"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/google_gemini_models"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "x-archimate-source": 1133
      }
    },
    "/api/google_gemini_models/{id}": {
      "get": {
        "operationId": "get_google_gemini_models",
        "summary": "Get GoogleGeminiModels by ID",
        "tags": [
          "google_gemini_models"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "operationId": "update_google_gemini_models",
        "summary": "Update GoogleGeminiModels",
        "tags": [
          "google_gemini_models"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/google_gemini_models"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "operationId": "delete_google_gemini_models",
        "summary": "Delete GoogleGeminiModels",
        "tags": [
          "google_gemini_models"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          }
        }
      }
    },
...
    "/api/stripe_webhook_handlers": {
      "get": {
        "operationId": "list_stripe_webhook_handlers",
        "summary": "List StripeWebhookHandlers",
        "tags": [
          "stripe_webhook_handler"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-archimate-source": 8763
      },
      "post": {
        "operationId": "create_stripe_webhook_handler",
        "summary": "Create StripeWebhookHandler",
        "tags": [
          "stripe_webhook_handler"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/stripe_webhook_handler"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "x-archimate-source": 8763
      }
    },
    "/api/stripe_webhook_handlers/{id}": {
      "get": {
        "operationId": "get_stripe_webhook_handler",
        "summary": "Get StripeWebhookHandler by ID",
        "tags": [
          "stripe_webhook_handler"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "operationId": "update_stripe_webhook_handler",
        "summary": "Update StripeWebhookHandler",
        "tags": [
          "stripe_webhook_handler"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/stripe_webhook_handler"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "operationId": "delete_stripe_webhook_handler",
        "summary": "Delete StripeWebhookHandler",
        "tags": [
          "stripe_webhook_handler"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "google_gemini_models": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "maxLength": 255
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "archived"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "status"
        ]
      },
...
      "stripe_webhook_handler": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "maxLength": 255
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "archived"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "status"
        ]
      },
...
    }
  }
}